crc32.c

+++ src/obfs/crc32.h
@@ -2,17 +2,18 @@
 #define _OBFS_CRC32_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 void init_crc32_table(void);
 
-uint32_t crc32_imp(unsigned char *buffer, unsigned int size);
+uint32_t crc32_imp(unsigned char *buffer, size_t size);
 
-void fillcrc32to(unsigned char *buffer, unsigned int size, unsigned char *outbuffer);
+void fillcrc32to(unsigned char *buffer, size_t size, unsigned char *outbuffer);
 
-void fillcrc32(unsigned char *buffer, unsigned int size);
+void fillcrc32(unsigned char *buffer, size_t size);
 
-void filladler32(unsigned char *buffer, unsigned int size);
+void filladler32(unsigned char *buffer, size_t size);
 
-int checkadler32(unsigned char *buffer, unsigned int size);
+int checkadler32(unsigned char *buffer, size_t size);
 
 #endif // _OBFS_CRC32_H
